+++ src/library/adash-errors.adb
@@ -52,6 +52,7 @@ package body Adash.Errors is
             | Error_No_History_Here
             | Error_History_Not_Forgotten
             | Error_Empty_Pipeline
+            | Error_Pipeline_Given_Up
             | Error_Too_Many_Kept
             | Error_Command_Bad_Assignment
             | Error_Script_Cycle
@@ -453,6 +454,8 @@ package body Adash.Errors is
          when Error_History_Not_Forgotten  =>
             return Adash.Messages.Msg_History_Not_Forgotten;
          when Error_Empty_Pipeline         => return Adash.Messages.Msg_Empty_Pipeline;
+         when Error_Pipeline_Given_Up      =>
+            return Adash.Messages.Msg_Pipeline_Given_Up;
          when Error_Too_Many_Kept          => return Adash.Messages.Msg_Too_Many_Kept;
          when Error_Command_Bad_Assignment => return Adash.Messages.Msg_Command_Bad_Assignment;
          when Error_Script_Cycle           => return Adash.Messages.Msg_Script_Cycle;